Add REST API setup for the disk selection
In guided setup we have additional steps in case multiple disks are
available, so user can select which disks to use, select which disk use
for the root partition and what to do with the existing partitions.

Adding initial structure to handle this step of guided setup.

See [poo#89932](https://progress.opensuse.org/issues/89932).