Increase time to wait for password prompt in ha-cluster-join
Changes introduced in crmsh include a 120s wait time in the cluster
bootstrap scripts for some scenarios (more than 3 cluster nodes, for
example). Current test code waits at most 60s for the second password
prompt when setting up a cluster with more than 3 nodes which is not
enough if this prompt takes longer to be shown.