Fix service check's console services ncurses font issue at bashrc file
On ppc64le, sometimes the console font will be distorted into pseudo graphics characters.
We need to reset the console font to fix it. As it impacted all the console services, added this command to bashrc file
Before it only put the code to fix kdump problem. Now all the console service has this problem, put this command to bashrc file, so it can fix all console font problem.