boot_to_desktop: close activities screen of GNOME 40+
Since GNOME 40 the user is presented with an open 'activities' screen
after login. The idea is that 'after starting the computer, the user
will want to start an application to actually use the computer'

The premisse is correct, but for openQA we want it 'easier' for us.
So, when we detect a booted desktop with activities open (based on
'gnome-activities' tag, we send ESC and close the overview.