Configure load_jeos_tests scheduling to run ltp tests in opensuse
`unless (get_var('LTP_COMMAND_FILE'))` is removed as it was blocking the
scheduling. ltp must have LTP_COMMAND_FILE and INSTALL_LTP.

`load_jeos_tests` was forcing `bootloader_uefi` to get scheduled twice.
i didnt find any reason for that but i didnt change it as there were objections
as the scheduling as it is, helps to debug failures.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>