Collect logs when linuxrc installation error banner
When the installation fails with "An error occurred during the installation."
during await_install a needle matches with tag `linuxrc_error` and dies immediately.
The linuxrc logs are stored in the *-y2log.tar.gz so no need for extra action in
the `post_fail_hook`.

I also refactored the module wrapping the statements that modify the timeout into a
subroutine to make the run more readable.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>