Replaces get_cmd_output in the integration_services_check_ip
Due to the random timeout in the connection which is used to run `run_ssh_cmd`,
we had to make sure that the connection is up or recreated.
os-autoinst has already merged the corresponding commits[0][1] that makes this possible,
as the paramemeter was not able to be passed through.
After the changes though the tests pass without the need to force the creation of a new
connection.
So we replace only the `get_cmd_output` which is deprecated with the `run_cmd` directly,
which with it turns calls `run_ssh_cmd` with `keep_open`. But keep_open is hold optional
for now.

[0] https://github.com/os-autoinst/os-autoinst/pull/1643/files
[1] https://github.com/os-autoinst/os-autoinst/pull/1642/files

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>