Set YUI REST API parameters for ssh installations
Libyui REST API linuxrc parameters we specify in the bootloader
are not passed to the installer, when it's started in the ssh, so we
should specify them in the same way, as we do when start YaST modules in
the installed system.
Besides explicit ssh installations, we also get to the ssh when perform
textmode installation on powerVM, zVM and zKVM. So this code will be
helpful in those cases too.