samba_adcli: Rework tests for joining a domain
Prior to this changes, we had some problems with the stability of this
test (#12417, #12394, #12386), due to a possibe bug in the adcli
utility (https://bugs.freedesktop.org/show_bug.cgi?id=55487 could be a
wild guess).

While there's room for improvement, this is already a stepping stone
into having more robust samba/AD/SSSD testing.

The test also should work on Tumbleweed but atm we lack the
infraestructure there to enable it for now