Implements docker firewall validation
I added a small validation for the expected configuration of the firewall.
Briefly, makes sure that Docker firewall zone exists with an active docker0
interface and in turn this is not added on trusted zone. Run a container and
verify the iptables.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>

Applies registry_url for the $registry assignment

registry_url to set the REGISTRY variable or docker.io wherever it is needed.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>