Add setters for host and port for YuiRestClient
For cases when network is restarted during test execution and IP address
of SUT is changed, YuiRestClient needs to know the new IP address to be
able to communicate with the SUT on remote installations.

The commit adds the appropriate setters for host and port, so that the
host can be changed at any time.