Fix AutoYaST mini in PowerVM
We were pre-formatting the disk at the same time that AutoYaST is
running and once the Beta warning dissapears there is nothing to
stop AutoYaST to continue and create a planning for partitioning
before we finish the wiping out process.

In general the logic to pre-format works for other scenarios so
we just need to keep the disk intact in this case and modify the
profile to wipe out the disk.