clamav: Check daily.cld if daily.cvd is not found
2. FreshClam checks the ClamAV virus database directory (indicated by the
DatabaseDirectory value in the freshclam.conf that FreshClam uses) for existing
instances of main.cvd, daily.cvd, or bytecode.cvd. For main and daily,
if the CVD can’t be found it also looks for main.cld and daily.cld.
These CLD files are uncompressed and unsigned versions of the CVD
that have had CDIFFs applied.

https://blog.clamav.net/2021/03/clamav-cvds-cdiffs-and-magic-behind.html
https://github.com/openmediavault/openmediavault/issues/702