Move remaining functions in partitions_validator_utils
In order to leave in partitions_validator_utils only functions that
are called by modules scheduled in maintenance group and not in
YaST group we can move this couple of functions to its respective
test module, as they are only used there.

In other words, check fstab or subvolume is out of scope here but should
be included in proper library in the future, for now is the way to break
the dependency with this non-general library in YaST group.