Fix mdmonitor.service failure
Until now an older mdadm version was used for systemd testsuite runs.
However this version is not available for 15-SP3.
The newer mdadm versions can only monitor if there is a Raid array
configured, so this commit adds this functionality, creating a Raid
array in order for mdmonitor.service to be running.