lib/transactional: Verify exit status in trup_call
While trup_call has code for waiting for the exit status, it didn't do much:
- It waited only for the expected exit status, so a mismatch would time out
- Mismatches were ignored completely

This commit changes it so that trup_call waits for transactional-update to
finish and then succeeds only exit status is as expected.

This is only possible when removing the code for the "$check == 0" case, but
that doesn't appear to be necessary: Even with rebootmgr doing an instant
reboot, the exit status reaches the serial console. This is not guaranteed by
the code (rebootmgrd calls systemctl reboot, both are async), but appears to
work in practice.