Improve failure reporting in verify_license_translations
Empty box was shown in openQA when the test module is failed during
license translations verifying.

The commit fixes the issue by removing EULA text from the failure box, as
it took too much space on failures and openQA seems to be not able to
process such huge messages in the failure box.

Related ticket: https://progress.opensuse.org/issues/92443