Fix grub_test_snapshot default snap was not shown
In grub_test_snapshot, if there are lots of snapshots, after rebooting
into snapshot we'll have lots of boot items in the grub boot list. so
the assert_screen 'default-snap' can't find the one with * before the
boot item. We need to move the cursor down to that item till we found it.