Containers: fix ip route command for all the SLE versions
In some SLE versions, the output of "ip r s default" shows not only
the default route, but also all the routes. For example, in SP1 it
returns:

ip r s default
default via 10.0.2.2 dev eth0 proto dhcp
10.0.2.0/24 dev eth0 proto kernel scope link src 10.0.2.15

but in SP3, it only returns the first line.

With this grep, we make sure we only output the first line.