publiccloud: improve stability of log_instance.sh
Before that patch log_instance.sh has intention to
support scenarios when it is triggered several times.
So PID file might contain PIDs of several processes
needs to be killed on tear down.
After examine current codebase there was no use case
for this feature. So it is simplified in favor of more
stable version when we expecting single PID in that file.
Another improvement is that from now on we checking if PID file
exists before deleting it.
Also there was decision to drop double-quotes during saving
PID into file.