Makes validate_btrfs module independent from the other container modules
This intends to fix the issue with the `/root/sle_base_image/docker_build`.
This used to work because the scheduling included the steps of the
directory's creation in another module which was running prior to the
`validate_btrfs`, which in turn, it reused it.

i have modified the module to become independent and atomic. That should
include the docker installation, configuration
(as it happens in docker.pm for example), and the creation and
removal of the resources.

In additional the test requires NUMDISKS eq to 2

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>