Containers: disable firewall
In some versions, we have firewall running and in some others firewall
is disabled. This way we make sure we start the test with the same
conditions for all the cases.

The reason to choose disabled is to avoid some further configuration
of iptables to allow containers contact the public network (bsc#1187176)
Since container tests are not targetting network testing, it makes
sense to run the general tests with additional network setup.
What we can do in the other hand is to create a new test case where
we test container networking with firewall enabled.