Handle both cases of "Yes/Trust" for unknown repo cert
Depending on content in the repositories sometimes it is needed to
confirm the unknown repo certificates, e.g. for NVidia, with "Yes",
sometimes it is "Trust", so try both variant key presses in rapid
succession, expecting that the second key press is not accepted
elsewhere on the same screen.

Related progress issue: https://progress.opensuse.org/issues/94534