Adds s390x autoyast profile for containers jobs
This PR creates a dedicate autoyast profile for s390 which will be used
in the containers group.
I havent touched the template profile `autoyast_containers.xml.ep`. This
still has conditions for s390 but it does not work.
The new autoyast profile is moved in a intiutive location under `autoyast_sle15/containers`.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>