wicked-wlan: Skip tests by wicked version
Introduces a check_wicked_version() method to verify if a test or config
is suitable to run (functionality implemented in wicked). This became
useful, because we would like to run this testsuite against different
wicked versions.

To skip a hole test module, it is possible to specify the wicked version
via a member variable like:
```
has wicked_version => '>=0.6.66'
```
This variable is checked in the wicked::wlan::run() method by default.
And if the current installed wicked version is to low, the test-module
get skipped.

If only one config of a test-module doesn't apply to the installed
wicked version, you can specify the config as a hash-ref like:
```
{
   wicked_version => '>=0.6.66',
   config => 'WIRELESS=yes'
}
```
And the default wicked::wlan::run() method will use this info to skip a
specific config or not.