Draft: OBS package build environment test
The main task is to 'co' (checkout) and 'build' locally a 'package' from
an 'API'/'codestream' using 'osc'. This is achieved by preparing the
systems with repositories and/or packages necessary, deploying and
modifying our configuration file and running the actual test. In order
to talk to the APIs, credentials must be provided as arguments.

It is running as root user as a normal user requires authentication, but
this could be improved. The 'testpkg' used, does not require an actual
build, but it does create an rpm (AC1: Any test executes 'osc build' on
our usual tested products). Included in the commit are:

  - [x] template oscrc file
  - [x] schedule changes

Related ticket: https://progress.opensuse.org/issues/36562
Needles: No Needles
Verification runs: TBA