Rename wmp_basic_test module to wmp_check_process
Updated documentation on WMP tests identifies 4 main tests:

- Do SAP Processes stay in cgroup? (former Basic Tests)
- Does memory.low cause performance penalty? (former Does WMP break
anything?)
- Does memory.low protect against swapping? (former Does WMP work?)
- HA Tests: Do SAP processes stay in cgroup?

As can be seen, WMP Basic Test used to refer to the test "Do SAP
Processes stay in cgroup?", but current test code for WMP in
os-autoinst-distri-opensuse is primarily used for the scenario
"HA Tests: Do SAP processes stay in cgroup?" in an HA+SAP scenario.

Moreover, the sles4sap/wmp_basic_test test module is just a piece in
testing the "HA Tests: Do SAP processed stay in cgroup?" scenario, as
the HA & SAP product configuration as well as the WMP setup is done
by other test modules; wmp_basic_test just checks that the processes
stay in the cgroup, so it makes sense to rename it to a name that better
reflects its current purpose.

This commit renames the test module to sles4sap/wmp_check_process.