Apply runtime.pm in containers libraries
Some hybrid approach to integrate the runtime.pm into the container tests.
This intends to improve a bit the design of the tests regarding containers.
In theory it gives you the possibility to initialize a runtime and implement
specifics of that particular Class instead of if/else inside the test.
For now the decoupling is not in place though.

- Use runtime.pm and its derived subclasses
- Some things added to runtime.pm but not everything is modified to avoid huge changes for now
- no change in the behavior or in the test logic

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>