toolkits: Explicitly set scaling factor to fix text rendering (boo#1137203)
A problem is that depending on the virtual graphics adapter, e.g. in our case
aarch64, the X server reports a zero sized screen and consequently a
practically infinite screen resolution.

Tk uses this information to initialize the scaling factor for translating
between points and pixels which in turn is used for rendering fonts if their
size is specified in points. This value can be set explicitly with "tk scaling
1.0" to ensure a proper text rendering.

Related bug: https://bugzilla.suse.com/show_bug.cgi?id=1137203