Fix free of run_ltp when instance restarts
On instance reboot, runltp hangs. The reason is that the restart script
forks a helper script for getting the logs, which keeps looping in the
background and perl waits for it to terminate. Since the helper-script
runs elsewhere it is not needed in the restart procedure we can simply
remove it from there.