boot_ltp: Take framebuffer console screenshot after boot
This is debug code for poo#81142. It'll gzip the contents of console
framebuffer and upload them to OpenQA.

The framebuffer contains raw pixel data of what the kernel thinks the screen
output should look like. The raw screenshot will not rule out a display device
driver bug in the kernel but it will tell us whether the bug is in the higher
level console rendering code.