Move buildah and validate_btrfs on host related tests
Find and reassign some modules to the correct group and run them only in a single place.

The reason they removed from image testing jobs are:

- buildah is unlikely to find any problem on the images other than podman will find.
this is because it used by podman underneath. Buildah is more likely to find bugs against
the sut rather the images.

- validate_btrfs tests the docker storage fs feature which likely belongs to the host tests as well.

HOST_VERSION is used in image-based tests and this was preventing to run buildah_docker in the
host-based tests.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>