Wait before checking screen at pxe grub boot
Due to the unique mechanism that is being used by openQA needle
matching program, even partially same areas in one needle can
not be matched as expected. And it is difficult to control the
search area by "area margin", so the optimal way is to delay the
needle matching by waiting for the the screen to stop change and,
at the same time, linux and initrd are already downloaded successfully.
Addtionally, use type_string_slow instead of type_string to better
cope slow response issue.