Fix host_configuration module against rhel
Running dhclient on all-good networking of the rhel image is messing
with the route table.

Running dhclient shouldnt cause any problem on an already configured route table
IIRC and is propably distro specific issue and it can ignored for now and workaround
exclude the execution when HDD is rhel image.

Maybe it would be nice to have an self-agnostic method here and run config
commands when it is needed.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>