Fix build_and_run_image for non-suse hosts
The `build_and_run_image` fails due to the Dockerfile context which
requires suse repos available but the base images are not supported
outside of the suse ecosystem.

The solution in this commit restores the Dockerfile that used to use
python image and no package installation was necessary. Use this when
host is not suse.

Extra modifications appears for rhel. as we removed the host_configuration
module, the select_console is required.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>