Fix qdevice/qnetd tests for diskless sbd scenario
HA qdevice/qnetd test modules are currently working on block device
backed sbd scenarios with no issues, however when using diskless sbd,
the test fails due to several issues:

1. When adding a qdevice in block device backed sbd scenarios, the
changes take effect automatically by reloading the cluster configuration
without a need to restart the cluster.

2. When adding a qdevice in diskless sbd scenarios, for changes to take
effect the cluster requires a restart. Bootstrap scripts will do this
automatically, but not when there are resources already configured in
the cluster, such as the one added by the ha/qnetd test module.

Due to this, it is necessary to restart the cluster services in both
nodes after adding back the qdevice in the diskless sbd scenario. This
commit adds code and barriers to sync the nodes and do this.

Additionally, after fencing the node in the diskless sbd with qdevice
scenario, the cluster is not coming back online due to bsc#1189398. This
commit also adds a workaround this issue, and a soft failure referencing
the bug.