Remove LV with dmsetup remove before lvremove
There is a sporadic issue with sles4sap/hana_install in MACHINE
64bit-ipmi-nvdimm in osd which is tracked in poo#96833. This commit
attempts to fix it by removing DM structures earlier with dmsetup remove
before any calls to lvremove, vgremove and pvremove.

In cases where this is not enough, the test will fail with a reference
to the poo ticket, and will attempt to gather more details during the
failure to further debug the issue.