Fix condition which assigns the proper dockerfile template
With the condition as it was before the `$dockerfile` variable was always
assigned to the one that uses the python image. That may not make the tests
fail but it was not the expected behavior.

The `dockerfile.python3` should be used only for the hosts which do not support
`container-suseconnect-zypp`. Use `dockerfile.python3` to workaround wherever we
cant perform apache installation as it happens in `Dockerfile`.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>