Add reset_console to test_users_locale sub when is not sle15
`glibc_locale.pm` was not written with sle12 in mind. When we scheduled
sle12sp5 we had an unexpected misbehavior with the console. Specifically
it makes `curl_https` break. The later switch to user console which left
dirty from `glibc_locale` as it exits when `is_sle('<15')`. When it is sle15
runs `test_users_locale` sub for a second time and then it resets the console
before exits.

With this patch we reset the consoles before the modules is completed even for
sle12

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>