Close Firefox by closing all tabs or killing it
Firefox will nag us when we try to close it via alt-f4 whether we want to close
all currently open tabs. The resulting pop up is however hard to distinguish
from a hanging Firefox. Instead, we just try to close all tabs via ctrl-w until
we see the desktop. If we don't get to see the desktop, then firefox is probably
hanging and we execute a kill -9 to just get rid of it.

Co-authored-by: Oliver Kurz <okurz@suse.de>