Add test summary with version information to hana_install
This commit adds a 'record_info()' call to the sles4sap/hana_install
test module that prints some information regarding the OS & HANA
versions used in the test to make it easier for reviewers to access
this information. Also included in the commit is the upload of
files containing the list of installed OS packages in the system and
the status of systemd services right after the HANA installation.

An 'HDB version' command was also added to the sles4sap/hana_test test
module to also supply HANA version information as well in this module.