Containers: remove BCI tests due to scheduling limitations
The addition of this new test is supposed to be executed only
for Container-Image-Updates, not for Server-DVD-Updates.
The problem is that the yaml schedule already has one condition
and we can't have double conditions in yaml... so let's just
remove it for now since it's breaking a lot of tests in maintenance.