Handle VERSION invocation in ensure_ca_certificates_suse_installed
In containers we use HOST_VERSION to identify the version that the host
is using and distinguish from the version that the container is running on.

`ensure_ca_certificates_suse_installed` omits the HOST_VERSION resulting
to use wrong repo to install ca-certifications in the host.

keep `get_required_var` to not change the current behavior.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>

Revert "Import gpg untrusted key changed to yes/no"

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>