Implement Factory pattern on containers with runargs as singleton
Extends runtime.pm to use a factory class for the creation of the
container runtime. With this design the divided modules of _podman and _docker
which have common code can be merged into one.

Also instead of Singleton class we can reuse the RunArgs to pass the runtime instance
to all the modules of the test.
Added TEST_CONTEXT to the yaml for now, make sure that works and it can be moved to the
job groups eventually

This implementation is based on https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/12842

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>