Remove reload_needles for set_var
In version_switch_upgrade_target.pm and version_switch_origin_system.pm
we call set_var with reload_needles => 1, which will reload all the needle
and it's a time consuming action. this will cause timeout when system boot
and resulted in boot into default grub item.

We can remove this to save time and will not effect the tests, because all
needles are not classfied with version.