Containers: Simplify containers schedule into single file
The reason for this proposal is to be able to follow in a
easier way the complex support matrix we have in containers
of different packages depending on sle versions and archs.
Currently, YAML schedule has a limitation in conditional
scheduling and this tries PR to solve that.

With this, we also remove all the duplication we have
currently in the yaml schedules and keep the conditions
in a single place.

Also, for HOST tests I grouped all the tests in a single job.
It will take longer time to execute but we keep all the tests
in a single place (this can be discussed).

The only downside is that we need to add `EXTRATEST=cont` to all
the container tests (it's already present in maintenance jobs)
and also `CONTAINER_RUNTIME` to help the scheduler pick docker
or podman tests.