Update design to use same module with both podman and docker
Those files have identical code. Using the Factory class we can
initialize the `$engine` without define which one it will be used.
I propose to define `PODMAN` and `DOCKER` althought we have already `CONTAINER_RUNTIME`. However
it becomes sort of deorecatd in this implementation.

Also when i started designing this i thought that the TestArgs class can be used to make the initialation
just once in the beginning. For the case where the same module is scheduled twice we can leverage the TestArgs
instance to indicate the desired container engine.

The idea here is to avoid to make repeatable calls in some cases, as some variables are common
under the same test job.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>