Containers: skip zypper related host tests for 15-SP4
Using zypper in containers with different version as the host
doesn't work if we are registered against SCC Proxy, which
gets the assets from local storage.

To avoid this, we skip the tests on latest products that are
not released (15-SP4) for this case and we set the variable
SCC_URL pointing to the corresponding proxy url.