containers::engine: Croak on timeout in _engine_script_run()
The function `testapi::script_run()` it self, does not have a proper
handling if the command didn't reach the end in the given timeout. For
this case it simply return `undef`. I guess all our use cases in
container doesn't need this flexibility, so we can take care of that
error.

The if condition for `$args{timeout} > 0` is needed, as `timeout => 0`
is used to execute a command, without waiting for the prompt.