containers/docker_image, containers/podman_image: Avoid testing 3rd party image
Currently it uses "Dockerfile.python3" unless on released SLE, but that one
always uses "docker.io/library/python:3" instead of the provided base. This
makes it unsuitable for the base image tests.

Instead, use "Dockerfile" which actually uses the provided base image if
possible, and don't build anything otherwise.

Note: This leaves the "Dockerfile.python3" code unused, it should be moved
to one of the container runtime tests.