BCI Containers: split main test module into 3 functional modules
To help troubleshooting and future enhancement of the tests,
splitting it into 3 different parts makes everything easier.
In OpenQA it will be shown which module fails and will be
quicker to spot the failure.
This also opens the door to easy integrate support for 3rd
party hosts in bci_prepare.pm