Allow external testkit to be supplied via a tarball
First version of console/systemd_sapstart_check expected the external
SAP testkit from a repository under the package name
SAP-systemdlib-tests, but stakeholders requested to drop this middle
step of package generation and publication into the repository. This
commit adapts the module so it can receive the tarball directly in the
QA_TEST_REPO setting, and extract the test kit from there before
running. To supply a tarball on the QA_TEST_REPO setting, it has to
point to the tarball location via a protocol that can be handled by
curl.