Update yamls which still use the old module names
For all those yamls we have to adapt the new names of the modules.
In addition we need to make sure that we update the job variables adding `CONTAINER_RUNTIME`
and separate each job for docker/podman.

Thats is because the `loadtest` in *scheduler.pm* is not flexible enough to handle running
both engines in the same job, as it is implemented currently.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>