[jeos] Design yamls with podman and docker to run as separate jobs
We need to split the jobs which get scheduled with both container engines together, to run separately.
Redesign the yamls schedules which fit those conditions to be able to do so.
This follows the latest trend in the qa-c.

The need cames from another task which tries to merge podman_image and docker_image in one side and
podman__3rd_party_images and docker_3rd_party_images.

Make the change in scheduling only for jeos for now to easy the followup work

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>