console/sshd: Improve background calls for port forwarding
The ssh calls in the background spam the serial terminal with debug messages,
which confuses serial input functions, leading to wait_serial not matching
and wasting a minute of waiting for each mismatch.

Write the debug output into log files and display it at the end instead.
Also make use of background_script_run.