Create named subroutines for test loader
That was one out of two things i left out of https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/13534

This is to make the scheduler readable as the else block in the `load_publiccloud_tests` was not self-explanatory.
i include also loader in replacement of `create_hdd_autoyast_pc.yaml` from the `PC Tools Image`.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>