Add installing state to test for Plasma PackageKit applet
Before this the test clicked the "Install Updates" button, then looked
at the screen waiting for needles that signal that the applet has
finished installing the updates.

This is bad because the needles, as currently captured, match while
the applet is still installing updates.

This commit fixes that by adding a check to make sure that when the
needles match the finished state, the SUT is not still installing
updates. If it is, the test will restart looking for the needles
matching the end of the installation. If it was the test will continue
on.

See: https://progress.opensuse.org/issues/101881