Disable junit report in qa_automation on SLE 15-SP4 and newer
Fix poo#102506: SLE 15-SP4 doesn't have python2 module. Junit report
script is python2 only. It was decided that old qa_automation scripts
will not be used and will be replaced in the future. It means that
script migration to python3 is not worth. We can still use
qa_automation scripts, but without human friendly openQA presentation.