Containers: specify swap partition size for HDD creation
If size is  not specified, the installation will assign 50% for swap
and 50% for root partition, which is not very efficient and we found
problems in tests complaining about no space left on the device.
The idea with this change is to limit the swap partition to 2GB
only and the rest will go to the root partition.