Do not run ltp_main when ltp test in public cloud
Using the same variables with the kernel tests creates some conflicts in the scheduling.
Specifically, the `LTP_COMMAND_FILE` indicates a ltp test. kernel is using it to schedule their
test, and is checked early in the main_common.

In this commit the check will ignore the kernel ltp if the test indicates public_cloud test.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>