Fix disks selection in Clone partition
There was a fix of an issue in libyui for YMultiSelectionBox.
Previously, checkboxes may be selected with both 'select' and 'check'
actions, that was incorrect. After the fix, 'select' action highlights
element in list and 'check' action checks a checkbox. Link to the PR
with fix: https://github.com/libyui/libyui/pull/58

The commit updates testing framework to use 'check' action to select
disks in 'Clone partition layout' dialog.