BCI: use retry feature from tox
From the Docs:
Some of the tests can be a bit flaky due to network resources
not being available. To avoid these issues, we make use of the
pytest-rerunfailures plugin. To enable it,
invoke tox with the --reruns command line flag as follows:
$ tox -e test_name -- --reruns 3 --reruns-delay 10
The option --reruns-delay delays the rerun (in this case) by 10 seconds,
thereby reducing the likelihood of another network issue.