Remove the record soft failure of bsc#1180080
Bug bsc#1180080 has been fixed, this pr is going to remove the soft failure
record. In migration cases we set cdrom as the first boot device to simplify
our migration progress. However, we need to boot the hard disk to fully
patch it before migration progress. For UEFI secure boot setup, the UEFI
signing key of the older system may not be up to date as newer IOS boot
image before fully patched, then boot failure if we try to boot the hard
disk from cdrom boot menu entry. So we keep this logic to boot the hard
disk directly.