Fix double interval while using wait_until
The interval was doubled while using wait_until function. It happened
because http requests already wrapped with wait_until function. In case
wait_until is used to wait for some events happened to widgets, the
interval that is specified in this wait_until become summarized with the
interval, that is used in WidgetController to send http requests.

The commit makes wait until to be dependent on time, but not on counter.
So, if it is set to wait two seconds, it will check current time, and
wait for two seconds from it, instead of just using counter increment.

Another change that helped to avoid double timeout is removal of "sleep"
from http_get and http_post. That was hard coded to wait 1 second for
each HTTP request.