Wait for keyboard layout to be selected on page load
When Product Selection Page is loaded, it takes some time for "English
(US)" to be selected. If during that period of time, another layout is
selected via LibyuiClient, it is not recognized by the system.

The commit ensures, "English (US)" is selected and then proceeds with
the further steps in the test.