Merge podman and docker image based tests into one module
Leverage `TestArgs` to pass an attribute which defines which container engine should be initialized for each case.
Because each engine has separate function which load the module, each `TestArgs` is also separate instance which means
that the modules works also when they both scheduled in the same job.

Because of the dependency of the current modules by JeOS yaml sceduling, those files were not removed as for now.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>