Handle polkit-agent-helper-1 not setuid root in CC
In cc setup, 'polkit-agent-helper-1' file is not setuid root
by default, then it will impact our polkit tests, current
workaround is "chmod u+s" for this file in our openQA test code.
it is a not good way to change the default file permission to
pass the test, so enhance our test code to handle it.